Refactoring: move inner classes to util package
